buildchain: add dependency checking for Dockerfiles
In order to track Dockerfile `COPY`/`ADD` image dependencies and
to correctly rebuild images when their dependencies change, we require
Dockerfile file dependencies to be tracked in the buildchain.

The `LocalImage` target parses its Dockerfiles, checks the parsed dependencies
against the local tree, and fails if dependencies are created in Dockerfiles
but not reflected in target file dependencies.

Issue: GH-885