Use "comment" keyword instead of "msg"
When voluntarily interrupting a formula, we use the
"test.fail_without_changes" state. To pass extra information, we used
the "msg" keyword, which is not valid. Instead, we now use the "comment"
keyword.

This was detected when the formulas could not find available IPs, and
the state return dict didn't show the expected information message.

Issue: GH-889